TST: Test eval_code_async (#1109)
